FT_RASTER_WARNING_DELTAC_IN_GLYPH_PGM
"Repeated executions in glyph programs can have unexpected results."
The DELTAC instruction changes the value of CVT entries. If this is executed in a glyph program, every time that glyph is hinted the CVT entry is changed. These changes will accumulate and cause a result that the hinter might not have expected.